+++ packages/tailwindcss/src/candidate.test.ts
@@ -1416,3 +1416,130 @@ it('should parse candidates with a prefix', () => {
     ]
   `)
 })
+
+it.each([
+  // Empty arbitrary value
+  'bg-[]',
+  'bg-()',
+  // — Tricking the parser with a space is not allowed
+  'bg-[_]',
+  'bg-(_)',
+
+  // Empty arbitrary value, with typehint
+  'bg-[color:]',
+  'bg-(color:)',
+  // — Tricking the parser with a space is not allowed
+  'bg-[color:_]',
+  'bg-(color:_)',
+
+  // Empty arbitrary modifier
+  'bg-red-500/[]',
+  'bg-red-500/()',
+  // — Tricking the parser with a space is not allowed
+  'bg-red-500/[_]',
+  'bg-red-500/(_)',
+
+  // Empty arbitrary modifier for arbitrary properties
+  '[color:red]/[]',
+  '[color:red]/()',
+  // — Tricking the parser with a space is not allowed
+  '[color:red]/[_]',
+  '[color:red]/(_)',
+
+  // Empty arbitrary value and modifier
+  'bg-[]/[]',
+  'bg-()/[]',
+  'bg-[]/()',
+  'bg-()/()',
+  // — Tricking the parser with a space is not allowed
+  'bg-[_]/[]',
+  'bg-(_)/[]',
+  'bg-[_]/()',
+  'bg-(_)/()',
+  'bg-[]/[_]',
+  'bg-()/[_]',
+  'bg-[]/(_)',
+  'bg-()/(_)',
+  'bg-[_]/[_]',
+  'bg-(_)/[_]',
+  'bg-[_]/(_)',
+  'bg-(_)/(_)',
+
+  // Functional variants
+  // Empty arbitrary value in variant
+  'data-[]:flex',
+  'data-():flex',
+  // — Tricking the parser with a space is not allowed
+  'data-[_]:flex',
+  'data-(_):flex',
+
+  // Empty arbitrary modifier in variant
+  'data-foo/[]:flex',
+  'data-foo/():flex',
+  // — Tricking the parser with a space is not allowed
+  'data-foo/[_]:flex',
+  'data-foo/(_):flex',
+
+  // Empty arbitrary value and modifier in variant
+  'data-[]/[]:flex',
+  'data-()/[]:flex',
+  'data-[]/():flex',
+  'data-()/():flex',
+  // — Tricking the parser with a space is not allowed
+  'data-[_]/[]:flex',
+  'data-(_)/[]:flex',
+  'data-[_]/():flex',
+  'data-(_)/():flex',
+  'data-[]/[_]:flex',
+  'data-()/[_]:flex',
+  'data-[]/(_):flex',
+  'data-()/(_):flex',
+  'data-[_]/[_]:flex',
+  'data-(_)/[_]:flex',
+  'data-[_]/(_):flex',
+  'data-(_)/(_):flex',
+
+  // Compound variants
+  // Empty arbitrary value in variant
+  'group-data-[]:flex',
+  'group-data-():flex',
+  // — Tricking the parser with a space is not allowed
+  'group-data-[_]:flex',
+  'group-data-(_):flex',
+
+  // Empty arbitrary modifier in variant
+  'group-data-foo/[]:flex',
+  'group-data-foo/():flex',
+  // — Tricking the parser with a space is not allowed
+  'group-data-foo/[_]:flex',
+  'group-data-foo/(_):flex',
+
+  // Empty arbitrary value and modifier in variant
+  'group-data-[]/[]:flex',
+  'group-data-()/[]:flex',
+  'group-data-[]/():flex',
+  'group-data-()/():flex',
+  // — Tricking the parser with a space is not allowed
+  'group-data-[_]/[]:flex',
+  'group-data-(_)/[]:flex',
+  'group-data-[_]/():flex',
+  'group-data-(_)/():flex',
+  'group-data-[]/[_]:flex',
+  'group-data-()/[_]:flex',
+  'group-data-[]/(_):flex',
+  'group-data-()/(_):flex',
+  'group-data-[_]/[_]:flex',
+  'group-data-(_)/[_]:flex',
+  'group-data-[_]/(_):flex',
+  'group-data-(_)/(_):flex',
+])('should not parse invalid empty arbitrary values: %s', (rawCandidate) => {
+  let utilities = new Utilities()
+  utilities.static('flex', () => [])
+  utilities.functional('bg', () => [])
+
+  let variants = new Variants()
+  variants.functional('data', () => {})
+  variants.compound('group', Compounds.StyleRules, () => {})
+
+  expect(run(rawCandidate, { utilities, variants })).toEqual([])
+})